Implemented VFS using http://jamesmolloy.co.uk/tutorial_html/8.-The%20VFS%20and%20the%20initrd.html just to get going

+++ libc/include/sys/types.h
@@ -10,7 +10,7 @@ typedef uint8_t dev_t;
 
 #ifndef _TYPEDEF_ino_t
 #define _TYPEDEF_ino_t
-typedef uint16_t ino_t;
+typedef uint32_t ino_t;
 #endif
 
 #ifndef _TYPEDEF_off_t
